[step2] feat: view, controller, application 생성
