+++ src/main/kotlin/blackjack/BlackjackApplication.kt
@@ -1,0 +1,7 @@
+package blackjack
+
+import blackjack.controller.BlackjackController
+
+fun main() {
+    BlackjackController.start()
+}
