+++ src/main/kotlin/blackjack/service/BlackJackService.kt
@@ -1,0 +1,7 @@
+package blackjack.service
+
+class BlackJackService {
+    fun splitInput(input: String): List<String> {
+        return input.split(",")
+    }
+}
