+++ src/test/kotlin/blackjack/domain/CardTest.kt
@@ -1,0 +1,12 @@
+package blackjack.domain
+
+import io.kotest.core.spec.style.StringSpec
+import io.kotest.matchers.shouldBe
+
+class CardTest : StringSpec({
+    "카드를 생성한다." {
+        val card = Card()
+        card.createCard()
+        card.cardList.size shouldBe 52
+    }
+})
